Add libifconfig module
I managed to find this in an old local git branch, and with some minor
tweaks it builds!  So here it is, pretty much as I found it.  I expect
there are parts of the API that need improvement, left for future work.

Big drawback, libifconfig(3) is private so we need to build it all into
this module out of the source tree.  It would be good to push
libifconfig(3) across the finish line and make it public so we can link
to a shared library instead.