sysctl: Fix logic to stop iterating
Iteration needs to stop when the next oid prefix no longer matches what
we started with.  However, this comparison was using the previous oid
length (a count of integers) rather than the prefix size (a count of
bytes).  Correct this by adding a mib field for the original prefix size
and using this for the comparison.  The mib.all field is eliminated as
it is now redundant.  A zero-length prefix is always matched, so we get
unimpeded iteration for free by setting the prefix size to zero.