sysctl: Include starting kern node mib when iterating all mibs
Normally, mib iteration only visits descendants.  This means skipping
the first (parent) mib.  However, iterating all mibs is a special case
where we want to include the starting mib.

This behavior is divergent from sysctl(8), e.g. sysctl -da will show a
description for vm but not for kern.  I consider this a bug in
sysctl(8).