libfetch: Make optional arguments optional
Be consistent with the rest of the library and make flags optional for
fetch.request(), along with the arguments that follow it.  They are all
optional for the C library.

Update the manpage to reflect the change.  While at it, make the
sendfile(3lua) documentation more idiomatic about how it distinguishes
multiple optional arguments.