sysctl: Avoid realloc
Simplify buffer allocation to merely free the old buffer and allocate a
new one when a value does not fit in the previously allocated buffer.
We have to read the whole value anyway, so there is no point preserving
the previous contents of the buffer.

While reorganizing this loop, there is no need to round up the size to a
multiple of PAGE_SIZE.  This is a short-lived, single-use buffer.  The
contents are immediately copied into Lua and then the buffer is freed.
We have no alignment enforced either.  Simply doubling the size is
enough to handle value growth.