README: Note that organization needs improvement
The module boundaries are not necessarily final, and could be improved
so that instead of a bunch of different modules for each function in
e.g. unistd.h, there is one libc module with submodules like
libc.unistd, libc.sys.stat, etc.  This will take some time to feel out
how to ultimately approach it.