Combine various modules into a freebsd module
Organize libc/librt bindings under a freebsd module divided into
submodules by header.  This somewhat overlaps with libfreebsd, which
serves a similar purpose.  There is so little in libfreebsd at the
moment that it is practical to reimplement the overlapping modules where
needed.

Adapt Makefiles to centralize the flua quirks into a top-level
Makefile.inc and include this from subdir makefiles.  This also fixes
building submodules from within the subdir instead of always needing to
build from the top of the repo.