Drop the top level freebsd module
On second thought, conflicting with the existing freebsd module name
is more trouble than it's worth.  Instead, drop the freebsd prefix.
This should avoid conflicts in the base system, though conflicts in the
broader Lua ecosystem may become apparent later when this project is
adapted to build for ports.  Punt for now.